Apply single responsibility to GenderController
- Moving validation rules to GenderRequest
- Cleaning uploading image to use upload_base64_image

+++ app/Http/Controllers/Api/Admin/Course/GenderController.php
@@ -5,9 +5,10 @@ namespace App\Http\Controllers\Api\Admin\Course;
 use Illuminate\Http\Request;
 use App\Models\Course\Gender;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Validator;
-use App\Http\Resources\Admin\Course\NamingResource;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
+use App\Http\Requests\Admin\Course\GenderRequest;
+use App\Http\Resources\Admin\Course\NamingResource;
 
 class GenderController extends Controller
 {
@@ -22,8 +23,8 @@ class GenderController extends Controller
         $genders = Gender::filter($request)->sortData($request)->offset($request->perPage * $request->page)->paginate($request->perPage);
 
         return response()->json([
-          'total'   => Gender::all()->count(),
-          'namings' => NamingResource::collection($genders)
+            'total'   => Gender::all()->count(),
+            'namings' => NamingResource::collection($genders)
         ]);
     }
 
@@ -40,42 +41,22 @@ class GenderController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  GenderRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(GenderRequest $request)
     {
-        // Validation
-        $validator = Validator::make($request->all(), [
-          'name_ar' => 'required|min:3|unique:genders',
-          'name_en' => 'nullable|min:3',
-          'description_ar' => 'nullable|min:3',
-          'description_en' => 'nullable|min:3'
-        ]);
-
-        if ($validator->fails()) {
-          return response()->json($validator->errors(), 400);
-        }
-
         // Upload Image
         if ($request->image) {
-          $base64Image  = explode(";base64,", $request->image);
-          $explodeImage = explode("image/", $base64Image[0]);
-          $imageGender    = $explodeImage[1];
-          $image_base64 = base64_decode($base64Image[1]);
-          $imageName    = uniqid() . '.'.$imageGender;
-          Storage::disk('public')->put("courses/namings/images/{$imageName}", $image_base64);
-          $request->merge(['image' => $imageName]);
+            $name = upload_base64_image($request->image, "courses/namings/images");
+            $request->merge(['image' => $name]);
         }
-
         // Store in database
         $gender = Gender::create($request->all());
-
         return response()->json([
-          'message' => __('messages.successful_create'),
-          'subscriber' => new NamingResource($gender)
+            'message' => __('messages.successful_create'),
+            'subscriber' => new NamingResource($gender)
         ], 200);
-
     }
 
     /**
@@ -86,56 +67,39 @@ class GenderController extends Controller
      */
     public function show(Gender $gender)
     {
-      return new NamingResource($gender);
+        return new NamingResource($gender);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  GenderRequest  $request
      * @param  Gender  $gender
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Gender $gender)
+    public function update(GenderRequest $request, Gender $gender)
     {
-        // Validation
-        $validator = Validator::make($request->all(), [
-          'name_ar' => 'required|min:3|unique:genders,name_ar,' . $gender->id,
-          'name_en' => 'nullable|min:3',
-          'description_ar' => 'nullable|min:3',
-          'description_en' => 'nullable|min:3'
-        ]);
 
-        if ($validator->fails()) {
-          return response()->json($validator->errors(), 400);
+        if (!$request->image && $gender->image) { // Delete the previous image
+            Storage::disk('public')->delete("courses/namings/images/{$gender->image}");
         }
 
         // Upload Image
         if ($request->image) {
-          if (str_starts_with($request->image, 'data:image')) {
-            $base64Image  = explode(";base64,", $request->image);
-            $explodeImage = explode("image/", $base64Image[0]);
-            $imageGender    = $explodeImage[1];
-            $image_base64 = base64_decode($base64Image[1]);
-            $imageName    = uniqid() . '.'.$imageGender;
-            // Delete the previous image
-            Storage::disk('public')->delete("courses/namings/images/{$gender->image}");
-            // Save the new image
-            Storage::disk('public')->put("courses/namings/images/{$imageName}", $image_base64);
-            $request->merge(['image' => $imageName]);
-          } else {
-            $request->merge(['image' => $gender->image]);
-          }
-        } else if($gender->image) {// Delete the previous image
-          Storage::disk('public')->delete("courses/namings/images/{$gender->image}");
+            if (str_starts_with($request->image, 'data:image')) {
+                // Delete the previous image
+                Storage::disk('public')->delete("courses/namings/images/{$gender->image}");
+                $name = upload_base64_image($request->image, "courses/namings/images");
+                $request->merge(['image' => $name]);
+            }
         }
 
         // Store in database
         $gender->update($request->all());
 
         return response()->json([
-          'message' => __('messages.successful_update'),
-          'subscriber' => new NamingResource($gender)
+            'message' => __('messages.successful_update'),
+            'subscriber' => new NamingResource($gender)
         ], 200);
     }
 
@@ -150,7 +114,7 @@ class GenderController extends Controller
         $gender->status = !$gender->status;
         $gender->save();
         return response()->json([
-          'message' => __('messages.successful_update')
+            'message' => __('messages.successful_update')
         ], 200);
     }
 
@@ -169,7 +133,7 @@ class GenderController extends Controller
         // Delete database record
         $gender->delete();
         return response()->json([
-          'message' => __('messages.successful_delete')
+            'message' => __('messages.successful_delete')
         ], 200);
     }
 }
